cloud/services/ingestor/requirements.txt
paho-mqtt==1.6.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==1.10.14
redis==5.0.1
python-dotenv==1.0.1
